// client/src/components/item/__generated__/ItemList_query.graphql.ts
export type ItemList_query$data = {
  readonly clogItemSearch: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly barcode: string | null | undefined;
        readonly createdAt: string;
        readonly id: string;
        readonly name: string;
        readonly stockCount: number;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number | null | undefined;
  } | null | undefined;
  readonly " $fragmentType": "ItemList_query";
};
export type ItemList_query$key = {
  readonly " $data"?: ItemList_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"ItemList_query">;
};